refactor: Move `map_batches` up to `EagerExpr`
